Fixes #1
Stops sleeping state when home memory is cleared on attempt to path if bed is marked and villager is sleeping. The sleep behavior doesn't run when game is closed and reloaded with a villager in bed and is what normally cleared the sleep state when a bed is marked with a villager still in it.